remove psf_photometry function
